refactor get_function_definitions
